Change Timeout selection to string from integer
When a users selects "inf" in the timeout combo box on the Canvas page,
that entry immediately gets converted to an integer value of 300.

This commit delays the conversion of DRAW_TIME to integer until its
needed in the timeout test.  The timeout test is modified to check
for the string value.  The configuration save code is also altered to
avoid a conversion to integer.

 Changes to be committed:
    modified:   bCNC/CNCCanvas.py